requirements.txt
langchain==1.2.1
python-dotenv==1.2.1
langchain-text-splitters==1.1.0
faiss-cpu==1.13.2
sentence-transformers==5.2.0
numpy==2.4.0
Pillow==12.1.0
mypy==1.19.1
black==25.12.0
asyncio==4.0.
langchain-community==0.4.1
langchain-experimental==0.4.1
anthropic==0.84.0
langchain-huggingface==1.2.1
docling==2.79.0
pypdf==6.8.0